feat: eita lasquera
