+++ js/i18n/en/configuracao-campeonato.json
@@ -1,6 +1,6 @@
 {
     "TituloPagina": "Championship Configuration",
-    "Configuracoes": "Configurations",
+    "ConfigMenu": "Configurations",
     "Editar": "Edit informations",
     "Times": "Teams",
     "Exclusao": "Exclusion",
