+++ js/i18n/ptbr/configuracao-campeonato.json
@@ -1,6 +1,6 @@
 {
     "TituloPagina": "Configuração de Campeonato",
-    "Configuracoes": "Configurações",
+    "ConfigMenu": "Configurações",
     "Editar": "Editar informações",
     "Times": "Times",
     "Exclusao": "Exclusão",
