+++ pages/configuracao-campeonato.html
@@ -15,8 +15,8 @@
 	<componente-header></componente-header>
 	<main class="container">
 		<div class="row mb-4 mt-5 h-100">
-			<section class="col-10 col-md-3 config-menu bg-white p-4 rounded-5">
-				<h5 class="i18" key="Configuracoes">Configurações</h5>
+			<section class="col-10 col-md-3 config-menu mb-5 mb-md bg-white p-4 rounded-5">
+				<h5 class="i18" key="ConfigMenu">Configurações</h5>
 
 				<!-- Config Menu -->
 				<ul class="list-group list-group-flush config-menu-list" id="config-menu-list">
