+++ pages/pagina-campeonatos.html
@@ -78,37 +78,7 @@
                 </div>
             </div>
         </section>
-
-        <div class="d-flex d-md-none flex-column mt-5">
-            <div class="row mb-2 justify-content-center">
-                <ul class="list-group list-group-horizontal list-group-flush justify-content-center p-0">
-                    <li class="list-group-item border-0 opacity-75 bg-transparent"><a href="https://github.com/RoDaMo"><i class="bi bi-github fs-1"></i></a></li>
-                </ul>
-            
-            </div>
-  
-            <hr class="opacity-50 border rounded-pill m-0 bg-dark d-flex">
-  
-            <div class="row mb-auto mt-2 justify-content-center">
-                <div class="col-auto">
-                        <ul class="list-group list-group-horizontal list-group-flush">
-                        <li class="list-group-item border-0 opacity-75 me-5 bg-transparent"><a href="#" class="i18" key="Politica">Política</a></li>
-                        <li class="list-group-item border-0 opacity-75 bg-transparent"><a href="#" class="i18" key="Termos">Termos</a></li>
-                    </ul>
-                        <ul class="list-group list-group-horizontal list-group-flush">
-                        <li class="list-group-item border-0 opacity-75 me-5 bg-transparent"><a href="#" class="i18" key="Sobre">Sobre</a></li>
-                        <li class="list-group-item border-0 opacity-75 bg-transparent"><a href="#">Cookies</a></li>
-                    </ul>
-                </div>
-            </div>
-              
-  
-            <div class="row mt-4">
-                <div class="col">
-                    <p class="text-black text-center opacity-50">&#169; 2023 RoDaMo</p>
-                </div>
-            </div>
-        </div>
+        
         <div class="mt-4">
             <p class="text-danger" id="mensagem-erro"></p>
         </div>
