+++ pages/pagina-times.html
@@ -108,10 +108,10 @@
                                     <!-- Camisetas do time -->
                                     <ul class="splide__list">
                                         <li class="splide__slide">
-                                            <img class="img-fluid rounded-4 jerseys-img" src="https://s2.glbimg.com/et8WqfpqRZ9QUE2FRW7EMduS2Tg=/e.glbimg.com/og/ed/f/original/2021/02/19/ancientrelic.jpg" alt="C1">
+                                            <img class="img-fluid rounded-4 jerseys-img" src="https://w7.pngwing.com/pngs/129/898/png-transparent-jersey-nike-sleeve-adidas-shirt-soccer-jerseys-tshirt-blue-sport-thumbnail.png" alt="C1">
                                         </li>
                                         <li class="splide__slide">
-                                            <img class="img-fluid rounded-4 jerseys-img" src="https://avatars.githubusercontent.com/u/97525102?v=4" alt="C1">
+                                            <img class="img-fluid rounded-4 jerseys-img" src="https://i.pinimg.com/550x/77/ac/1b/77ac1bcc08a7f141a50a3d2902d1ee17.jpg" alt="C1">
                                         </li>
                                     </ul>
                                 </div>
@@ -126,14 +126,14 @@
                         </div>
                         <div data-lenis-prevent class="d-flex px-3 pt-3 ss-third-content w-auto">
                             <div class="container d-flex justify-content-center">
-                                <div class="row gy-4 row-cols-auto row-cols-lg-2 row-cols-md-3 row-cols-sm-2 row-cols-xl-2 row-cols-xxl-3 mb-4 ss-slider-2">
+                                <div id="champstuff" class="row gy-4 row-cols-auto row-cols-lg-2 row-cols-md-3 row-cols-sm-2 row-cols-xl-2 row-cols-xxl-3 mb-4 ss-slider-2">
                                     <div class="col">
                                         <a href="#" class="text-decoration-none">
                                             <div class="rounded-5 ss-championship-content d-flex flex-column m-auto p-2">
                                                 <div class="ss-championship-img-wrapper position-relative rounded-circle overflow-hidden m-auto mt-3">
-                                                    <img class="img-fluid position-absolute mw-100 h-100" src="https://icatcare.org/app/uploads/2018/07/Thinking-of-getting-a-cat.png" alt="">
+                                                    <img class="img-fluid position-absolute mw-100 h-100" src="https://upload.wikimedia.org/wikipedia/en/thumb/f/f5/UEFA_Champions_League.svg/1200px-UEFA_Champions_League.svg.png" alt="">
                                                 </div>
-                                                <p class="text-center mt-2 mb-3 fs-5 text-nowrap text-truncate d-inline-block m-auto ss-championship-texts">nome</p>
+                                                <p class="text-center mt-2 mb-3 fs-5 text-nowrap text-truncate d-inline-block m-auto ss-championship-texts">UEFA Champions League</p>
                                             </div>  
                                         </a>
                                     </div>
@@ -174,42 +174,7 @@
                 </div>
             </div>
         </section>
-
-        <!-- Camisas do Time / Campeonatos -->
-
-        <div class="d-flex d-md-none flex-column mt-5">
-            <div class="row mb-2 justify-content-center">
-                <ul class="list-group list-group-horizontal list-group-flush justify-content-center p-0">
-                    <li class="list-group-item border-0 opacity-75 bg-transparent"><a href="https://github.com/RoDaMo"><i class="bi bi-github fs-1"></i></a></li>
-                </ul>
-            
-            </div>
-  
-            <hr class="opacity-50 border rounded-pill m-0 bg-dark d-flex">
-  
-            <div class="row mb-auto mt-2 justify-content-center">
-                <div class="col-auto">
-                        <ul class="list-group list-group-horizontal list-group-flush">
-                        <li class="list-group-item border-0 opacity-75 me-5 bg-transparent"><a href="#" class="i18" key="Politica">Política</a></li>
-                        <li class="list-group-item border-0 opacity-75 bg-transparent"><a href="#" class="i18" key="Termos">Termos</a></li>
-                    </ul>
-                        <ul class="list-group list-group-horizontal list-group-flush">
-                        <li class="list-group-item border-0 opacity-75 me-5 bg-transparent"><a href="#" class="i18" key="Sobre">Sobre</a></li>
-                        <li class="list-group-item border-0 opacity-75 bg-transparent"><a href="#">Cookies</a></li>
-                    </ul>
-                </div>
-            </div>
-              
-  
-            <div class="row mt-4">
-                <div class="col">
-                    <p class="text-black text-center opacity-50">&#169; 2023 RoDaMo</p>
-                </div>
-            </div>
-        </div>
-        <div class="mt-4">
-            <p class="text-danger" id="mensagem-erro"></p>
-        </div>
+        
     </main>
 
     <componente-footer></componente-footer>
