+++ scss/pagina-times.scss
@@ -182,6 +182,8 @@ body, html {
     height: 5rem;
     min-width: 5rem;
 
+    background-color: $white;
+
     img {
         min-width: 100%;
     }
